chore: make it a feature
