+++ Cargo.toml
@@ -37,7 +37,8 @@ alloy = { version = "=0.5.4", default-features = false, features = ["consensus",
 revm = { version = "18.0.0", default-features = false }
 
 zenith-types = { version = "0.10", optional = true }
-dashmap = "6.1.0"
+
+dashmap = { version = "6.1.0", optional = true }
 
 [dev-dependencies]
 alloy-rlp = { version = "0.3", default-features = false }
@@ -57,6 +58,7 @@ tokio = { version = "1.39", features = ["macros", "rt-multi-thread"] }
 [features]
 default = [
     "std",
+    "concurrent-db",
     "revm/std",
     "revm/c-kzg",
     "revm/blst",
@@ -65,6 +67,8 @@ default = [
 ]
 
 std = ["revm/std", "alloy/std", "alloy-rlp/std", "alloy-primitives/std", "alloy-sol-types/std", "dep:zenith-types"]
+
+concurrent-db = ["std", "dep:dashmap"]
 
 test-utils = ["revm/test-utils", "revm/std", "revm/serde-json", "revm/alloydb"]
