+++ src/db/sync_state.rs
@@ -1,4 +1,5 @@
 use crate::db::ConcurrentCacheState;
+use alloc::{collections::BTreeMap, vec::Vec};
 use alloy_primitives::{Address, B256, U256};
 use dashmap::mapref::one::RefMut;
 use revm::{
@@ -9,10 +10,7 @@ use revm::{
     primitives::{Account, AccountInfo, Bytecode},
     Database, DatabaseCommit, DatabaseRef, TransitionAccount, TransitionState,
 };
-use std::{
-    collections::{hash_map, BTreeMap},
-    sync::RwLock,
-};
+use std::{collections::hash_map, sync::RwLock};
 
 /// State of blockchain.
 ///
